server: Port modified files and the remaining of msg.ts to rust

+++ server-rs/src/action_types.rs
@@ -824,7 +824,7 @@ pub struct IModifiedFilesChanged {
     pub scanning: bool,
     pub full: bool,
     pub changed: Vec<ModifiedFile>,
-    pub removed: Vec<f64>,
+    pub removed: Vec<i64>,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, TS)]
